Add test support for multiple connections in Http2
